+++ src/main.py
@@ -1,8 +1,6 @@
 import websocket, json
 from model import dglab_message
 from handler import *
-from PIL import ImageGrab
-from pyzbar.pyzbar import decode
 import pydglab
 import asyncio
 import logging
@@ -10,6 +8,7 @@ import yaml
 import time
 from rich.table import Table
 from rich.live import Live
+from utils import find_qr_code
 
 with open("./config.yaml") as stream:
     try:
@@ -23,14 +22,6 @@ formatter = logging.Formatter(
     "%(module)s [%(levelname)s]: %(message)s"
 )
 
-def capture_fullscreen():
-    screenshot = ImageGrab.grab()
-    return screenshot
-
-def decode_qr_code(image):
-    decoded_objects = decode(image)
-    qr_data = [obj.data.decode('utf-8') for obj in decoded_objects]
-    return qr_data
 if config["Debug"]:
     fh = logging.FileHandler("debug.log")
     fh.setLevel(logging.DEBUG)
@@ -49,16 +40,7 @@ while qrRaw == "":
 
     time.sleep(1.5)
 
-    qrcodes = decode_qr_code(capture_fullscreen())
-
-    if len(qrcodes) < 1:
-        logging.error("识别二维码失败")
-        continue
-
-    for i in range(len(qrcodes)):
-        if qrcodes[i].find("wss://") != -1 or qrcodes[i].find("ws://") != -1:
-            qrRaw = qrcodes[i]
-            break
+    qrRaw = find_qr_code()
 
     if qrRaw == "":
         logging.info("没有有效的 Websocket 二维码")
